requirements.txt
django==3.0.9
python-dotenv==0.14.0
django-admin-sortable2==0.7.7
django-tinymce==3.0.2